Checking key in addition to the 'key' property
The Jira REST API has a slight difference between
the cloud and server versions. The server version
doesn't have a 'key' value in the entrys in the
'fields' (it is a hash) property of the 'issuetype'
resource. This will check the actual key in the 'fields'
hash in the case that it doesn't have a value called 'key'